External context creation without callbacks
This change separates the concept of the external heap from the
external context, i.e., the context and the heap may be allocated
in separate memory regions by the user, and what's more, it puts
the allocation under the direct control of the user without any
need for a callback mechanism (opposed to the existing approach
based on `jerry_context_alloc_t` callbacks).

The here-implemented approach
- enables the same context/heap separation in the external case
  that is already available in the global internal context
  configuration, and
- enables statically allocated external heaps.

However, as this change gives allocation in the hands of the user,
the binding of the heap to the context can only happen when the
engine is initialised. This necessitates the change of the
arguments of `jerry_init`, which is a major API break (as that
function cannot be avoided in any scenario involving the engine.)

JerryScript-DCO-1.0-Signed-off-by: Akos Kiss akiss@inf.u-szeged.hu