Calculate string lengths only when needed
It is not necessary to calculate the length values for strings on
every operation. This patch aims to improve performance by delaying
the length computations until the value is actually required.
This patch was done in cooperation with Tamás Kéri.

JerryScript-DCO-1.0-Signed-off-by: Mátyás Mustoha mmatyas@inf.u-szeged.hu