Seperate tests from test-api.c - Strings
`test-api.c` is containing the testing for all API functions, which resulted in a big and hardly readable file.
This PR is the first of many, that tries to segment it into more files, by tested functionality, this time strings and string methods.

JerryScript-DCO-1.0-Signed-off-by: Bela Toth tbela@inf.u-szeged.hu