Add documentation about x86 and floating-point operations
If JerrScript is compiled for x86 then it is possible that the
floating-point operations are not correct for the JS standard.

Reproduction example:

```
$ ./tools/build.py --compile-flag=-m32  --compile-flag=-std=gnu99
$ echo "0.42018199893505187" | ./build/bin/jerry
jerry> 0.420181998935051836951970472000539302825927734463817841970...
```

To "fix" this the users should be advised to use at least SSE2
during build for this platform.

```
$ ./tools/build.py --compile-flag=-m32 --compile-flag=-std=gnu99 \
    --compile-flag=-mfpmath=sse --compile-flag=-msse2
$ echo "0.42018199893505187" | ./build/bin/jerry
jerry> 0.42018199893505187
```

JerryScript-DCO-1.0-Signed-off-by: Dániel Vince vinced@inf.u-szeged.hu